requirements.txt
tensorflow==2.10
pandas
numpy
pyyaml
gdown